+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Tag diagnostic HTTP messages with an internal ID, to make it easier to cross reference them.
 
 ## [0.41.0] - 2026-08-07
 ### Added